P4ADEV-1677 move PagoPaNodeFaults to enums package
