+++ src/main/java/it/gov/pagopa/pu/pagopapayments/service/synchronouspayments/PaymentService.java
@@ -12,7 +12,7 @@ import it.gov.pagopa.pu.pagopapayments.connector.auth.AuthnService;
 import it.gov.pagopa.pu.debtpositions.dto.generated.InstallmentDTO;
 import it.gov.pagopa.pu.pagopapayments.mapper.PaGetPaymentMapper;
 import it.gov.pagopa.pu.pagopapayments.mapper.PaVerifyPaymentNoticeMapper;
-import it.gov.pagopa.pu.pagopapayments.util.PagoPaNodeFaults;
+import it.gov.pagopa.pu.pagopapayments.enums.PagoPaNodeFaults;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang3.tuple.Pair;
 import org.springframework.stereotype.Service;
